[req-changes] Moved WiFi Mesh integration test to separate file
