+++ requirements-test.txt
@@ -6,3 +6,4 @@ freezegun~=1.1.0
 openwisp-monitoring @ https://github.com/openwisp/openwisp-monitoring/tarball/master
 openwisp-controller @ https://github.com/openwisp/openwisp-controller/tarball/master
 openwisp-utils[qa,selenium] @ https://github.com/openwisp/openwisp-utils/tarball/master
+redis~=4.5.4
